+++ lib/src/analyzers/lint_analyzer/anti_patterns/pattern_utils.dart
@@ -1,8 +1,6 @@
-import 'package:analyzer/dart/ast/ast.dart';
 import 'package:source_span/source_span.dart';
 
 import '../models/issue.dart';
-import '../models/scoped_function_declaration.dart';
 import 'models/pattern.dart';
 
 Issue createIssue({
@@ -31,17 +29,3 @@ Uri documentation(Pattern pattern) => Uri(
         pattern.id,
       ],
     );
-
-int getArgumentsCount(ScopedFunctionDeclaration dec) {
-  final declaration = dec.declaration;
-
-  int? argumentsCount;
-  if (declaration is FunctionDeclaration) {
-    argumentsCount =
-        declaration.functionExpression.parameters?.parameters.length;
-  } else if (declaration is MethodDeclaration) {
-    argumentsCount = declaration.parameters?.parameters.length;
-  }
-
-  return argumentsCount ?? 0;
-}